+++ src/Traversable.ts
@@ -27,28 +27,10 @@
  * @since 2.0.0
  */
 import { Applicative, Applicative1, Applicative2, Applicative2C, Applicative3, Applicative3C } from './Applicative'
-import {
-  Foldable,
-  Foldable1,
-  Foldable2,
-  Foldable2C,
-  Foldable3,
-  FoldableComposition,
-  FoldableComposition11,
-  getFoldableComposition
-} from './Foldable'
-import {
-  Functor,
-  Functor1,
-  Functor2,
-  Functor2C,
-  Functor3,
-  FunctorComposition,
-  FunctorComposition11,
-  getFunctorComposition
-} from './Functor'
-import { HKT, Kind, Kind2, Kind3, URIS, URIS2, URIS3 } from './HKT'
+import { Foldable, Foldable1, Foldable2, Foldable2C, Foldable3 } from './Foldable'
 import { pipe } from './function'
+import { Functor, Functor1, Functor2, Functor2C, Functor3 } from './Functor'
+import { HKT, Kind, Kind2, Kind3, URIS, URIS2, URIS3 } from './HKT'
 
 /**
  * @since 3.0.0
@@ -266,101 +248,27 @@ export interface Sequence3<T extends URIS3> {
 /**
  * @since 3.0.0
  */
-export interface TraversableComposition<F, G> extends FoldableComposition<F, G>, FunctorComposition<F, G> {
-  readonly traverse: <H>(
-    H: Applicative<H>
-  ) => <A, B>(f: (a: A) => HKT<H, B>) => (fga: HKT<F, HKT<G, A>>) => HKT<H, HKT<F, HKT<G, B>>>
-  readonly sequence: <H>(H: Applicative<H>) => <A>(fga: HKT<F, HKT<G, HKT<H, A>>>) => HKT<H, HKT<F, HKT<G, A>>>
+export function traverseComposition<F, G>(
+  F: Traversable<F>,
+  G: Traversable<G>
+): <H>(H: Applicative<H>) => <A, B>(f: (a: A) => HKT<H, B>) => (fga: HKT<F, HKT<G, A>>) => HKT<H, HKT<F, HKT<G, B>>> {
+  return (H) => {
+    const traverseF = F.traverse(H)
+    const traverseG = G.traverse(H)
+    return (f) => traverseF((ga) => pipe(ga, traverseG(f)))
+  }
 }
 
 /**
  * @since 3.0.0
  */
-export interface TraverseComposition11<F extends URIS, G extends URIS> {
-  <H extends URIS3>(H: Applicative3<H>): <R, E, A, B>(
-    f: (a: A) => Kind3<H, R, E, B>
-  ) => (fga: Kind<F, Kind<G, A>>) => Kind3<H, R, E, Kind<F, Kind<G, B>>>
-  <H extends URIS2>(H: Applicative2<H>): <E, A, B>(
-    f: (a: A) => Kind2<H, E, B>
-  ) => (fga: Kind<F, Kind<G, A>>) => Kind2<H, E, Kind<F, Kind<G, B>>>
-  <H extends URIS2, E>(H: Applicative2C<H, E>): <A, B>(
-    f: (a: A) => Kind2<H, E, B>
-  ) => (fga: Kind<F, Kind<G, A>>) => Kind2<H, E, Kind<F, Kind<G, B>>>
-  <H extends URIS>(H: Applicative1<H>): <A, B>(
-    f: (a: A) => Kind<H, B>
-  ) => (fga: Kind<F, Kind<G, A>>) => Kind<H, Kind<F, Kind<G, B>>>
-  <H>(H: Applicative<H>): <A, B>(f: (a: A) => HKT<H, B>) => (fga: Kind<F, Kind<G, A>>) => HKT<H, Kind<F, Kind<G, B>>>
-}
-
-/**
- * @since 3.0.0
- */
-export interface SequenceComposition11<F extends URIS, G extends URIS> {
-  <H extends URIS3>(H: Applicative3<H>): <R, E, A>(
-    fga: Kind<F, Kind<G, Kind3<H, R, E, A>>>
-  ) => Kind3<H, R, E, Kind<F, Kind<G, A>>>
-  <H extends URIS2>(H: Applicative2<H>): <E, A>(
-    fga: Kind<F, Kind<G, Kind2<H, E, A>>>
-  ) => Kind2<H, E, Kind<F, Kind<G, A>>>
-  <H extends URIS2, E>(H: Applicative2C<H, E>): <A>(
-    fga: Kind<F, Kind<G, Kind2<H, E, A>>>
-  ) => Kind2<H, E, Kind<F, Kind<G, A>>>
-  <H extends URIS>(H: Applicative1<H>): <A>(fga: Kind<F, Kind<G, Kind<H, A>>>) => Kind<H, Kind<F, Kind<G, A>>>
-  <H>(H: Applicative<H>): <A>(fga: Kind<F, Kind<G, HKT<H, A>>>) => HKT<H, Kind<F, Kind<G, A>>>
-}
-
-/**
- * @since 3.0.0
- */
-export interface TraversableComposition11<F extends URIS, G extends URIS>
-  extends FoldableComposition11<F, G>,
-    FunctorComposition11<F, G> {
-  readonly traverse: TraverseComposition11<F, G>
-  readonly sequence: SequenceComposition11<F, G>
-}
-
-/**
- * Returns the composition of two traversables
- *
- * @example
- * import { traversableArray } from 'fp-ts/lib/Array'
- * import { applicativeIO } from 'fp-ts/lib/IO'
- * import { none, traversableOption, some } from 'fp-ts/lib/Option'
- * import { getTraversableComposition } from 'fp-ts/lib/Traversable'
- *
- * const T = getTraversableComposition(traversableArray, traversableOption)
- * const state: Record<string, number | undefined> = {
- *   a: 1,
- *   b: 2
- * }
- * const read = (s: string) => () => state[s]
- * const x = T.sequence(applicativeIO)([some(read('a')), none, some(read('b')), some(read('c'))])
- * assert.deepStrictEqual(x(), [some(1), none, some(2), some(undefined)])
- *
- * @since 3.0.0
- */
-export function getTraversableComposition<F extends URIS, G extends URIS>(
-  F: Traversable1<F>,
-  G: Traversable1<G>
-): TraversableComposition11<F, G>
-export function getTraversableComposition<F, G>(F: Traversable<F>, G: Traversable<G>): TraversableComposition<F, G>
-export function getTraversableComposition<F, G>(F: Traversable<F>, G: Traversable<G>): TraversableComposition<F, G> {
-  const FuC = getFunctorComposition(F, G)
-  const FoC = getFoldableComposition(F, G)
-  return {
-    map: FuC.map,
-    reduce: FoC.reduce,
-    foldMap: FoC.foldMap,
-    reduceRight: FoC.reduceRight,
-    traverse: (H) => {
-      const traverseF = F.traverse(H)
-      const traverseG = G.traverse(H)
-      return (f) => traverseF((ga) => pipe(ga, traverseG(f)))
-    },
-    sequence: (H) => {
-      const sequenceF = F.sequence(H)
-      const sequenceG = G.sequence(H)
-      return (fgha) => sequenceF(pipe(fgha, F.map(sequenceG)))
-    }
+export function sequenceComposition<F, G>(
+  F: Traversable<F>,
+  G: Traversable<G>
+): <H>(H: Applicative<H>) => <A>(fga: HKT<F, HKT<G, HKT<H, A>>>) => HKT<H, HKT<F, HKT<G, A>>> {
+  return (H) => {
+    const sequenceF = F.sequence(H)
+    const sequenceG = G.sequence(H)
+    return (fgha) => sequenceF(pipe(fgha, F.map(sequenceG)))
   }
 }