+++ test/FunctorWithIndex.ts
@@ -1,11 +1,11 @@
 import * as assert from 'assert'
 import * as A from '../src/Array'
-import * as _ from '../src/FunctorWithIndex'
+import { mapWithIndexComposition } from '../src/FunctorWithIndex'
 import { pipe } from '../src/function'
 
 describe('FunctorWithIndex', () => {
-  it('getFunctorComposition', () => {
-    const F = _.getFunctorWithIndexComposition(A.functorWithIndexArray, A.functorWithIndexArray)
+  it('mapWithIndexComposition', () => {
+    const mapWithIndex = mapWithIndexComposition(A.functorWithIndexArray, A.functorWithIndexArray)
     const f = ([i, j]: readonly [number, number], a: string) => a + i + j
     assert.deepStrictEqual(
       pipe(
@@ -13,7 +13,7 @@ describe('FunctorWithIndex', () => {
           ['a', 'b'],
           ['c', 'd']
         ],
-        F.mapWithIndex(f)
+        mapWithIndex(f)
       ),
       [
         ['a00', 'b01'],